% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{add_ignore}
\alias{add_ignore}
\title{Add to ignore file}
\usage{
add_ignore(file, add, path = ".")
}
\arguments{
\item{file}{Either ".Rbuildignore" or ".gitignore"}

\item{add}{The lines to add}

\item{path}{Path to the package}
}
\value{
Returns TRUE invisibly
}
\description{
Add to ignore file
}
