% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{build_compile_commands}
\alias{build_compile_commands}
\title{Generate Compilation Database for Use with Clang Tools}
\usage{
build_compile_commands(path = ".", debug = FALSE)
}
\arguments{
\item{path}{The path of the package}

\item{debug}{Set to TRUE to get verbose output}
}
\description{
Generate Compilation Database for Use with Clang Tools
}
