% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{check_clang}
\alias{check_clang}
\title{Check Whether a Clang Toolchain is Being Used}
\usage{
check_clang()
}
\value{
A logical indicating whether clang is used as a compiler
}
\description{
Check Whether a Clang Toolchain is Being Used
}
