% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{has_clang}
\alias{has_clang}
\title{Check Whether a Clang Toolchain is Being Used}
\usage{
has_clang()
}
\value{
A logical indicating whether clang is used as a compiler
}
\description{
Check Whether a Clang Toolchain is Being Used
}
