% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{is_unix}
\alias{is_unix}
\title{Checks that We are on a Unix Platform}
\usage{
is_unix()
}
\value{
A logical indicating whether we are on a Unix platform
}
\description{
Checks that We are on a Unix Platform
}
